Moved transaction database calls into their own repository
