+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/Repositories/IDasLevyRepository.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using SFA.DAS.EAS.Domain.Data.Entities.Account;
 using SFA.DAS.EAS.Domain.Models.Levy;
 using SFA.DAS.EAS.Domain.Models.Payments;
-using SFA.DAS.EAS.Domain.Models.Transaction;
 
 namespace SFA.DAS.EAS.Domain.Data.Repositories
 {
@@ -17,12 +16,12 @@ namespace SFA.DAS.EAS.Domain.Data.Repositories
         Task<DasDeclaration> GetLastSubmissionForScheme(string empRef);
         Task<DasDeclaration> GetSubmissionByEmprefPayrollYearAndMonth(string empRef, string payrollYear, short payrollMonth);
         Task ProcessDeclarations();
-        Task<List<TransactionLine>> GetTransactionsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
-        Task<int> GetPreviousTransactionsCount(long accountId, DateTime fromDate);
+       
+     
         Task<List<AccountBalance>> GetAccountBalances(List<long> accountIds);
         Task CreateNewPeriodEnd(PeriodEnd periodEnd);
         Task<PeriodEnd> GetLatestPeriodEnd();
-        Task<List<TransactionLine>> GetTransactionDetailsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
+     
         Task CreatePaymentData(PaymentDetails paymentDetails);
         Task<Payment> GetPaymentData(Guid paymentId);
         Task ProcessPaymentData();
